# Add cache of configuration files list