Performance result:

command          | before       |  after
--------------------- | ------------- | ---------
setup:uninstall   | 3s             |  0.9 s
setup:install   |  32s          |  16s