@danslo 2015-08-08
+1 on this. Natively on linux (on an SSD) there is no problem, but when installing via a vagrant box that gets files over a share, there's a lot of this:

<img src="http://i.imgur.com/9rmxyM6.png" />

And that repeats over and over.