@vancoz 2015-08-11
Hi @otakarmare, this approach has no cache invalidation, and Reader is questionable place to implement this caching, also it looks inconstant with other methods - getComposerJsonFiles, getActionFiles