@vancoz 2015-08-11
Hi @otakarmare, cache invalidation should not be an issue, have a question about getComposerJsonFiles, getActionFiles methods, probably is it good idea to add caching for them too?