@danslo 2015-08-11
Indeed cache validation doesn't seem like an issue because this cache is per-request.