@otakarmare 2015-08-12
@vancoz I agree that getComposerJsonFiles, getActionFiles copy/paste part of logic with getConfigurationFiles and Magento/Framework/Module/Dir/Reader class has too many responsibilities, but it's another issue and I do not plan to do it in this pull request