@otakarmare 2015-10-05
I sign CLA
any updates from magento team?