@vrann 2015-10-07
Can you please fix Travis builds?