@Vinai 2015-11-03
@otakarmare Can you please rebase onto the current develop branch and sign the licence cla agreement? Thanks!