@mazhalai 2016-02-18
@otakarmare can you please merge the latest develop and restart builds?