@mazhalai 2016-02-23
Will be merged with MAGETWO-49470